requirements.txt
openai
Telethon
tinkoff-investments
tqdm
schedule
pytz